Add files via upload
Added the beginnings of an invincibility system whereby ships can't take damage right after being hit. While doing this, I found out that damage taken is not only proportional to the number of beams to intersect with the ship hull but also proportional to length of beam that intersects with the hull, due to the way the damage detector works. I don't know if I want to keep it or get rid of it.