Figured out a secret
Found out that dividing two integers by each other will always result in an integer. Modified some functions in physics.py to take this into account which helps reduce the error in the angle aligning function in general_ai.py but the error is still significant and scales with the frame rate which is the latest mystery.